model-2-bertopic/requirements.txt
bertopic==0.16.2
joblib==1.4.2
numba==0.58.1
sentence-transformers==3.0.1
scikit-learn==1.0.2
umap-learn==0.5.3
hdbscan==0.8.29
flask
cython 
pandas
